Fix message + remove zip
