+++ containers/ecr-viewer/docker-compose.yaml
@@ -2,8 +2,6 @@ name: ecr-viewer
 
 services:
   ecr-viewer:
-    ports:
-      - "3000:3000"
     build:
       context: ./
       dockerfile: Dockerfile
